requirements.txt
Django==1.11.4
django-bootstrap-form==3.3
djangorestframework==3.7.0
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytz==2017.2
